+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+**Version: 2.0.1**
+
+	* Added an helpers file.
+	* format_number() is no longer a method in the cart library, it is now a function inside the helpers file.
+	* Renamed the cart add() method to insert()
+	* Added in a few Exceptions.
+	* Some code improvements.
+	* Updated the examples.
+	* Added the assets folder, missing in the previous version.
+
 **Version: 2.0.0**
 
     * Changed bundle name from Lara-Cart to Cartify